refactor: move experience root to the root of the folder
