--- packages/experience-builder-sdk/src/blocks/index.ts
@@ -1,1 +1,0 @@
-export { ExperienceRoot } from './ExperienceRoot';
